Reading Files
Add a file menu option to the main menu in your TODO list app. In this
file menu create a (O)pen file option. The user should be able to enter
the path to a file to have it loaded.

Create a function that reads the file, and adds each line to the
current TODO list. Loading data/todo_list.txt should properly load the
list from above. Be sure to fclose() the file when you are done reading
it.